roll_error/requirements.txt
antspyx==0.4.2
cmocean==2.0
dipy==1.7.0
itk==5.3.0
itk_core==5.3.0
itk_elastix==0.17.1
itk_filtering==5.3.0
itk_io==5.3.0
itk_numerics==5.3.0
itk_registration==5.3.0
itk_segmentation==5.3.0
matplotlib==3.5.3
numpy==1.21.6
obspy==1.3.1
opencv_python==4.8.1.78
osgeo==0.0.1
pandas==1.3.5
psutil==5.9.6
pyelastix==1.2
rasterio==1.2.10
scikit_image==0.19.3
scikit_learn==0.22.1
scipy==1.7.3
spectral==0.23.1
torch==1.13.1
tqdm==4.66.1
